# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API que gerencia os usuários, as reservas, os espaços e os pagamentos",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/space": {
            "get": {
                "description": "Retorna um array com todos os espaço existentes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spaces"
                ],
                "summary": "Lista todos os espaços",
                "responses": {
                    "200": {
                        "description": "Lista de espaços existentes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Space"
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum espaço encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um espaço com os dados enviados no corpo da requisição. Apenas administradores podem criar espaços.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spaces"
                ],
                "summary": "Cria um novo espaço",
                "parameters": [
                    {
                        "description": "Dados do novo espaço",
                        "name": "space",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.RequestSpace"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Espaço criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/models.Space"
                        }
                    },
                    "400": {
                        "description": "Usuário não encontrado ou não é administrador",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Espaço já existe",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/space/{id}": {
            "get": {
                "description": "Retorna um JSON com os dados do espaço especificado pelo ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spaces"
                ],
                "summary": "Retorna um espaço específico",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do espaço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espaço encontrado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/models.Space"
                        }
                    },
                    "404": {
                        "description": "Espaço não encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de um espaço existente com base no ID informado. Apenas administradores podem atualizar espaços.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spaces"
                ],
                "summary": "Atualiza um espaço existente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do espaço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados atualizados do espaço",
                        "name": "space",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.RequestSpace"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Espaço atualizado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/models.Space"
                        }
                    },
                    "400": {
                        "description": "Usuário não encontrado, não é admin, senha incorreta ou espaço não existe",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deleta um espaço existente pelo ID informado. Apenas administradores podem deletar espaços.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Spaces"
                ],
                "summary": "Deleta um espaço específico",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do espaço",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Informações do usuário (username e password)",
                        "name": "space",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/space.RequestSpace"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Espaço deletado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Senha incorreta ou dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Espaço não encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Retorna um array com todos os usuários cadastrados",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Lista todos os usuários",
                "responses": {
                    "200": {
                        "description": "Lista de usuários cadastrados",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "404": {
                        "description": "Nenhum usuário encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um usuário com username e password informados no corpo da requisição",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Cria um novo usuário",
                "parameters": [
                    {
                        "description": "Dados do novo usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.RequestUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "409": {
                        "description": "Usuário já existe",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "chave que aparece no JSON",
                    "type": "string"
                }
            }
        },
...
        "models.Space": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "location": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "substation": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "isAdmin": {
                    "type": "boolean"
                },
                "passHash": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "space.RequestSpace": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "png": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "price": {
                    "type": "integer"
                },
                "substation": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.RequestUser": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}